gdb.rocm/device-attach.exp: mark untested if ptrace attach not permited
A GNU/Linux system can be configured to restrict the ability to trace
a running process with ptrace (see /proc/sys/kernel/yama/ptrace_scope in
man ptrace(2)).

When such mechanism is enabled, the gdb.rocm/device-attach.exp testcases
fail.

This patch proposes to handle this situation gracefully by using the
gdb_attach / gdb_attach_spawn_cmdline TCL procs introduced recently.
This allows the tests to be marked UNSUPPORTED instead of having plain
FAIL errors.

While at updating this test file, do some refactoring in
process_attach_tests (which checks that we can attach to a process using
the 'attach' command from GDB's prompt).  The test really tests this
command twice, once with the objfile pre-loaded (using the 'file'
command) and once without the objfile pre-loaded.  Instead of having
multiple part of the code dealing with those scenarios, just iterate
along a "preloadfile" axis (i.e. on/off), and only pre-load the file in
the on iteration.  The rest of the test can be shared, and the list
threads / show a GPU thread's architecture part can be done in both
cases instead of just one previously.

If ptrace restrictions are enabled, running the test before this patch
gives:

    Running .../gdb/testsuite/gdb.rocm/device-attach.exp ...
    FAIL: gdb.rocm/device-attach.exp: process_attach_tests: detach, with file (the program is no longer running)
    FAIL: gdb.rocm/device-attach.exp: process_attach_tests: extract thread id
    FAIL: gdb.rocm/device-attach.exp: process_attach_tests: detach, without file (the program is no longer running)
    FAIL: gdb.rocm/device-attach.exp: process_attach_cmdline_tests: detach the process (the program is no longer running)

                    === gdb Summary ===

    # of expected passes            5
    # of unexpected failures        4

With this patch, it becomes:

    Running .../gdb/testsuite/gdb.rocm/device-attach.exp ...

                    === gdb Summary ===

    # of expected passes            1
    # of unsupported tests          3

Tested on x86_64-gnu-linux.

Change-Id: Id0ea66b888f06db626b8c2deb9e8fafd0440fa01