gdb/infrun: Allow new threads in handle_vfork_done
The commit d8bbae6ea08 "gdb: fix handling of vfork by multi-threaded
program (follow-fork-mode=parent, detach-on-fork=on)" did introduce a
regression in gdb.rocm/fork-exec-gpu-to-non-gpu.exp:

    (gdb) break break_here_execer
    Function "break_here_execer" not defined.
    Make breakpoint pending on future shared library load? (y or [n]) y
    Breakpoint 1 (break_here_execer) pending.
    (gdb) PASS: gdb.rocm/fork-exec-gpu-to-non-gpu.exp: detach-on-fork=on: follow-fork-mode=parent: fork_func=vfork: gdb_breakpoint: set breakpoint at break_here_execer
    run
    Starting program: /home/lancesix/src/rocm-gdb/_build/gdb/testsuite/outputs/gdb.rocm/fork-exec-gpu-to-non-gpu/fork-exec-gpu-to-non-gpu-execer-vfork
    [Thread debugging using libthread_db enabled]
    Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".
    [New Thread 0x7ffff6627700 (LWP 229798)]
    [New Thread 0x7fffe1fff700 (LWP 229799)]
    [New Thread 0x7ffff5d4e700 (LWP 229800)]
    [Thread 0x7fffe1fff700 (LWP 229799) exited]
    [Detaching after vfork from child process 229801]
    FAIL: gdb.rocm/fork-exec-gpu-to-non-gpu.exp: detach-on-fork=on: follow-fork-mode=parent: fork_func=vfork: runto: run to break_here_execer (timeout)
    continue
    FAIL: gdb.rocm/fork-exec-gpu-to-non-gpu.exp: detach-on-fork=on: follow-fork-mode=parent: fork_func=vfork: continue until exit at continue parent to end (timeout)

The issue is linked to the new implementation to handle VFORK.  Please
refer to d8bbae6ea08 for a discussion about the new handling strategy.

In non-stop mode, the new implementation needs to stop all threads in
the inferior and then allows only the leading thread to be resumed in
order to report the VFORK_DONE event:

    /* If handling a vfork, stop all the inferior's threads, they will be
       restarted when the vfork shared region is complete.  */
    if (tp->pending_follow.kind () == TARGET_WAITKIND_VFORKED
        && target_is_non_stop_p ())
      stop_all_threads ("handling vfork", tp->inf);

Doing this,  stop_all_threads makes sure that new threads cannot be
created on the inferior by calling:

     prevent_new_threads (true, nullptr, -1, reason);

The issue we have is that when we see the expected VFORK_DONE event in
non stop mode, we restore breakpoints and release all the threads, but
fail to allow the creation of new threads:

    if (target_is_non_stop_p ())
      {
        scoped_restore_current_thread restore_thread;
        insert_breakpoints ();
        restart_threads (event_thread, event_thread->inf);
        start_step_over ();
      }

In the failing testcase, the parent calls vfork and then tries to launch
waves on an AMDGPU.  But because new thread creation is disabled, no new
waves are scheduled on the hardware.  Therefore the expected breakpoint
hit never happens and it is never reported, leading to the test's
timeout and FAIL.

This commit fixes this issue by ensuring that new thread creation is
allowed after the call to restart_threads (unless if scheduler locking
is in action).

Since there is already one instance of a call to prevent_new_threads
after a call to restart_threads, remove it because it is now redundant.

One thing to note.  Since stop_all_threads prevent thread creation in
all inferiors (whatever the inf parameter), this commit does the same in
restart_threads.  However, I feel that both should only affect the
inferior passed as argument if any.  There was already a FIXME note
regarding this point in a call to prevent_new_threads, so I keep it.

Regression tested on x86_64-gnu-linux + AMDGPU gfx1030

Change-Id: I4949e09d06fee6d33c08d5073336ca627884af6b
(cherry picked from commit bac316eedd881e44f31bfc9b10b01eebc9251988)