amd-dbgapi-target::resume: Ignore signo if not for a GPU thread
Since the introduction of gdb.threads/next-fork-other-thread.exp (in
27f9f649753b8c4bd4c40bde0f49d916c222a16a "gdb: resume ongoing step after
handling fork or vfork") we have the following FAILs:

    Running gdb/testsuite/gdb.threads/next-fork-other-thread.exp ...
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=fork: target-non-stop=off: non-stop=off: displaced-stepping=auto: i=17: next to other line
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=fork: target-non-stop=off: non-stop=off: displaced-stepping=on: continue
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=fork: target-non-stop=off: non-stop=off: displaced-stepping=on: i=9: next to break here
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=fork: target-non-stop=off: non-stop=off: displaced-stepping=off: continue
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=vfork: target-non-stop=off: non-stop=off: displaced-stepping=auto: continue
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=vfork: target-non-stop=off: non-stop=off: displaced-stepping=on: continue
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=vfork: target-non-stop=off: non-stop=off: displaced-stepping=on: i=0: next to other line (GDB internal error)
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=vfork: target-non-stop=off: non-stop=off: displaced-stepping=off: continue
    FAIL: gdb.threads/next-fork-other-thread.exp: fork_func=vfork: target-non-stop=off: non-stop=off: displaced-stepping=off: i=0: next to other line (GDB internal error)

This is mostly due to the presence of this unexpected message in GDB's
output:

    Resuming with signal SIGCHLD is not supported by this agent.

To reproduce it, let's take this simple example:

    #include <sys/wait.h>
    #include <stdlib.h>
    #include <unistd.h>

    int main ()
    {
      pid_t pid = fork ();
      if (pid == 0)
        exit (0);

      int ws;
      waitpid (pid, &ws, 0);

      return 0;
    }

and debug it with:

    ./gdb/gdb -data-directory ./gdb/data-directory \
              -ex "maintenance set target-non-stop off" \
              ~/tmp/a.out \
              -ex "handle SIGCHLD stop pass"
              -ex "r" \
              -ex "c" \
              -batch

The issue appears in amd_dbgapi_target::resume.  In this case, resume is
called with:

  signo == GDB_SIGNAL_CHLD and ptid == [XXXXX, 0, 0]

This means that the thread which should receive the signal SIGNO is a
INFERIOR_PTID ([XXXXX, XXXXX, 0] for instance).  In the current
implementation, "beneath ()->resume (ptid, step, signo);" is called,
making sure that the linux_nat target transmits the signal SIGNO.  Once
this is done, all other threads needs to be resumed (incl GPU threads as
indicated by the value of ptid).

As a preamble, the amd_dbgapi_target tries to map SIGNO to an
amd_dbgapi_exceptions_t value, in case SIGNO is intended for a GPU
thread.  This fails with a call to "error" since SIGCHLD does not make
sense for GPUs.

Since the thread which should receive the signal is not a GPU thread,
computing this mapping is useless, even if the SIGNO value could have a
mapping.   The resulting amd_dbgapi_exceptions_t will never be used.

To fix the error, make sure to only map a SIGNO to a
amd_dbgapi_exceptions_t if the thread which should receive the signal is
a GPU thread.  That is if PTID is a GPU thread, or if PTID is a wild
card ptid and INFERIOR_PTID is a GPU thread.

While at it, this patch moves the SIGNO to amd_dbgapi_exceptions_t
conversion before the call to beneath ()->resume ().  This way, if the
target thread is indeed a GPU thread and SIGNO has no mapping, we error
out before resuming anything, being on the host or on the GPU.

Tested on x86_64-linux-gnu + gfx900 and 86_64-gnu-linux + gfx1030.

Change-Id: I1affafa342da29619486a57f0f09c88203b2d39f
(cherry picked from commit bfa3fbf3fa99d5c7244b422802efde473e0e5909)