Handle externalized static globals mangling scheme
On AMDGPU, static globals may actually have extern linkage.  In
addition, in separate compilation mode, i.e., with -fgpu-rdc, such
globals must each have their own linkage name, because different
translation units may have different static globals with the same
name.

To do that, LLVM appends a ".static.$HASH" suffix to the mangled name.
For example:

 # 0x00000048:   DW_TAG_variable
 #                DW_AT_name      ("static_global")
 #                DW_AT_type      (0x00000044 "int")
 #                DW_AT_external  (true)
 #                DW_AT_decl_file ("[...snip...]/static-globals.cpp")
 #                DW_AT_decl_line (23)
 #                DW_AT_location  (DW_OP_addr 0x3004, DW_OP_stack_value, DW_OP_deref_size 0x8, DW_OP_constu 0x0, DW_OP_LLVM_form_aspace_address)
 #                DW_AT_linkage_name      ("_ZL13static_global.static.e5e0dfb539fa610f")

This commit teaches libiberty to demangle such symbols.

Currently, GDB/libiberty can't demangle such a symbol:

 (gdb) demangle -l c++ _ZL13static_global.static.e5e0dfb539fa610f
 Can't demangle "_ZL13static_global.static.e5e0dfb539fa610f"

The change to d_encoding is necessary otherwise d_encoding enters the
'if' and ends up returning NULL when the encoding is followed by a
vendor suffix, even after having successfully parsed the encoding,
which explains the failure to demangle the symbol at all.

The rest of the changes are needed because otherwise this vendor
suffix is incorrectly interpreted as a clone suffix, resulting in a
demangled name like this, for the example above:

   "static_global [clone .static] [clone .e5e0dfb539fa610f]"

After this change, the demangled name is just "static_global".

The loop in cplus_demangle_mangled_name is tweaked such that
supporting multiple vendor extensions is easier, with each vendor
extension d_foo function being responsible for either parsing the
extension successfully or not advancing the string.

libiberty demangler testcase included.  This adds tests for handling
multiple suffixes, though I haven't seen such a symbol yet.  I imagine
that it should be possible for a static function to be cloned, and
then externalized.

No regressions on x86-64 GNU/Linux.

Change-Id: I314efff780afa1450a2a5fabca01be3bf776e7e8