Make DWARF evaluator be more context sensitive
Until now, the DWARF expression evaluator only verified the existence
of a frame and a compilation unit context (where appropriate).  The
idea is that any existing DWARF operator that requires a frame to be
present, also requires a thread to be in focus as well.  The exception
to this is the DW_OP_form_tls_address operation, where the check for a
focused thread was implicitly done by the call to a symbol_read_needs
from a read_var_value function.

With the introduction of address spaces, the need for ensuring that a
correct thread and/or SIMD lane context is present becomes more
necessary.

This change introduces a thread context check to the ensure_have_frame
function and adds the relevant checks to appropriate DWARF operations:
DW_OP_form_tls_address, DW_OP_LLVM_form_aspace_address,
DW_OP_LLVM_aspace_bregx and DW_OP_LLVM_push_lane.

Change-Id: Ia3a2772273c3fd27427b3b3bad34f6b2e2d88ac0