Add mem_addr_ranges to struct value closure API
To support watchpoints on struct value objects that are categorized as
computable (DWARF composites and implicit pointers for example), we
need a way to extract all address ranges (that could support
watchpoints) encapsulated by the closure of that struct value object.

Similarly, we also need a way to traverse the nested location
description organization, following the logic of previously defined
dwarf_location methods.

Unfortunately, the result of this traversal can't be cached for the
next query, because it is dependent on a starting bit offset and
given size.

Change-Id: Ib614d2ee69e28cb569c5e8698675513be1937696
(cherry picked from commit 09f1587271b9b9c35715c59316f256c7bb759a58)