Add test for watchpoints in default and private_lane address spaces
This change introduces a simple test for watchpoints in default and
private_lane address spaces.  The idea is that a variable array is
expected to be allocated in a private_lane memory and should also have
an alias address in the default address space (global memory).

There are a few scenarios that this test is trying to cover:
 - watchpoint on a variable name (array) in a private_lane address
   space that gets triggered in the original context (frame, thread and
   lane),
 - watchpoint on the expression that describes an address of a single
   array element that gets triggered in the same context,
 - watchpoint on a single array element address directly that gets
   triggered under the context the address is bound to,
 - watchpoint that gets triggered in the callee function for all three
   previous scenarios,
 - watchpoint on a valid private_lane address set in a callee function
   but expected to get triggered in a caller function because the
   context of that address is still valid,
 - watchpoint on a kernel argument that is a pointer to the default
   address space that gets triggered in both callee and caller
   functions and
 - watchpoint on an address in a default address space directly that
   gets triggered in both callee and caller functions.

Current restriction for more test scenarios is with the missing address
space information for the pointers to the generic address space.  The
idea would be to test if inserting a watchpoint on a value of a callee
argument directly, would still trigger from a write to the same address
on a caller side.

Change-Id: Id807ffb8bd0f1c2827f15cfd3ca7a37d40558f60
(cherry picked from commit 76cc12ad7ff12f38f8d4517215db5281cc959d25)