gdb/testsuite: Fix gdb.rocm/*spurious-waves.exp
On gfx1030 the tests gdb.rocm/*spurious-waves.exp are failing.  This is
because the "hog" process which is meant to use most but not all
resources from the GPU is not properly configured for this architecture.

This patch proposes to have the hog process compute on the fly the
amount of compute resources it will hold.

The heuristic might be tuned more in the future if required.

This patch also fixes gdb.rocm/multi-inferior-run-spurious-waves.exp.  The
test at some point tries to run a second inferior (while inferior 1 is still
running) and checks that the second inferior hits a breakpoint.  However, a
call to "delete_breakpoints" prevented any breakpoint to be inserted in
inferior 2.  Fix this issue by removing this delete_breakpoints call.

Tested on x86_64-linux + gfx1030 and x86_64-linux + gfx90a.

Change-Id: I324bc9c9d9b23b6ee2a78db5ecdd68494d389e5a