gdb: Add maint amdgpu info address-alias command
It can sometimes be useful to be able to convert address from an
address space to another one (if such conversion is possible).  This
patch proposes to add a maintenance command which shows for a given
address (in any address space) the aliasing addresses in other address
spaces.  It also show the number of consecutive bytes this conversion is
valid for.

Here are example outputs on this new command:

    (gdb) maintenance info amdgpu address-aliases &idx
    Address space                       Address        Valid for
    global                       0x7ffff58f5e00              0x4
    generic             generic#0x2000000000018       0xffffffe8
    private_wave             private_wave#0x600              0x4
    private_lane              private_lane#0x18       0xffffffe8
    local                       <not reachable>              0x0
    (gdb) maintenance info amdgpu address-aliases gidx_p
    Address space                       Address        Valid for
    global                      <not reachable>              0x0
    generic             generic#0x1000000000000      0x100000000
    private_wave                <not reachable>              0x0
    private_lane                <not reachable>              0x0
    local                             local#0x0      0x100000000

Tested on x86_64-linux + gfx900.

Change-Id: I9d118f208589c3db6a2ec966ac221b1a54d0ccb6