gdb: make regcache::raw_update switch to right inferior
I noticed this small bug while working on rocm-gdb.

Let's say we have
two inferiors with the following target stacks:

    inf 1            inf 2
    -----            -----
                     amd-dbgapi
    linux-nat        linux-nat
    exec             exec

With the following patch applied (which makes amd-dbgapi not pushed on
inferiors not using the ROCm runtime), but without the change made in
the current patch, I would see some failures with this error message:

  Couldn't get extended state status: No such process.

This is an error message output by the linux-nat target trying to read
the registers of an unknown process.

What happened was:

 - do_target_wait makes inferior 2 current, does a target_wait, which
   returns a stop event for an amd-dbgapi wave (thread).
 - fetch_inferior_event calls switch_to_target_no_thread with linux-nat
   as the process target, since linux-nat is officially the process
   target of inferior 2.  This makes inferior 1 the current inferior.
 - In handle_signal_stop, we have:

    ecs->event_thread->suspend.stop_pc
      = regcache_read_pc (get_thread_regcache (ecs->event_thread));

    context_switch (ecs);

   regcache_read_pc executes while inferior 1 is still the current one,
   although the regcache is for a ptid managed by the amd-dbgapi
   target (e.g. (12345, 1, 1)).  The fetch_registers target call goes
   directly to the linux-nat target, which gets confused.

GDB should switch to the inferior the ptid belongs to before doing the
target call, to make sure the target call hits the right target stack.
The question is, who should ensure the current inferior is the right
one?  Given the regcache has a ptid and a process stratum target, it is
able to find the inferior this ptid belongs to and switch to it
temporarily if needed.  This keeps the regcache interface independent of
the global context, allowing callers to not have to care about switching
to a particular inferior before doing an operation on a regcache.

So, make regcache::raw_update switch to the right inferior before
calling target_fetch_registers.

Change-Id: Ic46187b38e2a41508797ca67d9989b1c5bb4e98c