gdb: record "prevent new threads" state in inferior
A subsequent patch will make the amd-dbgapi target not always pushed on
native inferiors.  When it gets pushed, it needs a way to know the
current "prevent new threads" mode for that inferior and adjust itself
accordingly.  The "prevent new threads" mode is set when the user
resumes execution (based on the "set scheduler-locking" mode, among
other things).  The easiest way to do this is to save the value in the
inferior struct, in target_prevent_new_threads.

Clear the flag in target_pre_inferior, in case the previous execution of
the inferior ended with the flag set.

Change-Id: I0343c3531f5a36d3f688de0eb43cde59202da189