gdb/amd-dbgapi: push target only when runtime is loaded
We currently push the amd-dbgapi target on native inferiors as soon as
they start executing (or as soon as we attach).  This means the
amd-dbgapi call occupies the arch stratum slot and intercepts target
calls even if not really necessary, when the inferior doesn't use the
ROCm runtime.

This patch changes it so that the amd-dbgapi target only pushes itself
on an inferior's target stack when that inferior loads the ROCm runtime.
And it unpushes itself when the inferior unloads the runtime.

The reason we initially wanted the target always pushed was to properly
handle exec and fork, like transferring the "precise memory" property
from fork parent to fork child, even if the parent did not have the ROCm
runtime loaded.  By having the target always pushed, it could intercept
the follow_exec and follow_fork target calls even in inferiors that
don't have the runtime loaded.  It's possible to achieve the same thing
with observers.  I think it's preferable this way.  The amd-dbgapi
target is only "observing" on the side line as long as the inferior
doesn't use the ROCm runtime, and isn't interfering at all.  When the
inferior does load the runtime, then it jumps in the game and starts
intercepting calls.

The changes in the amd-dbgapi target are:

 - Make enable_amd_dbgapi not push the amd-dbgapi target.  We only
   make amd-dbgapi (the library) attach the process and give us a
   notifier fd for it.
 - In the notifier fd callback (dbgapi_notifier_handler), we distinguish
   the cases where:

     1. amd-dbgapi is not pushed in that inferior, so the event must be
        a "runtime loaded" event.  In this case, we consume the event
	and push the target.  Apply the "prevent new threads" state
	in effect for that inferior.
     2. amd-dbgapi is already pushed in that inferior, in which case it
        means the runtime has been loaded previously.  We mark the async
	event handler such that the event will be consumed and handled
	through inferior_event_handler.

 - Modify handling of AMD_DBGAPI_EVENT_KIND_RUNTIME in process_one_event:
   if we get there it's because the target is pushed, because the
   runtime was loaded successfully previously.  The only runtime
   transition that can happen here is LOADED_SUCCESS -> UNLOADED.
 - Make disable_amd_dbgapi expect that the target may not be pushed.
 - Make amd_dbgapi_target::prevent_new_threads not check if the inferior
   is attached by amd-dbgapi.  If the target is pushed, it implies that.
 - Remove follow_fork and follow_exec target methods, replace with
   equivalent observers.  The code remains the same.
 - Replace the code in amd_dbgapi_target::close, which resets the
   amd-dbgapi lib, to a new function "maybe_reset_amd_dbgapi" called
   by disable_amd_dbgapi.
 - Call disable_amd_dbgapi in a new inferior_pre_detach observer.  We need
   to amd-dbgapi-detach before we ptrace-detach.  Without this
   observer, we would only amd-dbgapi-detach in the inferior_exited
   observer, which would be too late.

Change-Id: Ib68ae30db2a8e34ab74cb794acf5f137de366742