gdb/testsuite: Relax regex for the info agents command
The way dbgapi gets agent name will change.  We are not guaranteed
anymore that an agent name will consist in a name (\w+).  Update the
regexs in the tests to accomodate for arbitrary agent names.

Also, new version of dbapi gets the agent names from the pci.ids
database.  The path to this database is configured at dbgapi build time.
If dbgapi is misconfigured or the execution environment does not
include the pci.ids file where dbgapi searches it, a warning will be
printed.  This patch looks for this warning and emits a fail in the
testsuite if one is seen.  The intent is to help catch dbgapi packaging
and installation errors which would show up during gdb execution.

Change-Id: Id6cd9b01791e46b8ad078da7d504588a38b94270